Delete Reducer folder and replace with indiual name to save the slices
En lugar de tener Reducers folder por cada slice que se crea va a tener una carpeta con su nombre

Fix the export default to export the reducer of the slice, formart the code
