+++ client/src/Redux/user/userSlice.js
@@ -1,0 +1,28 @@
+import { createSlice } from '@reduxjs/toolkit'
+
+const initialState = {
+  user: 'jhonson',
+  userError: null,
+  userLoading: false,
+  token: '',
+}
+
+const userSlice = createSlice({
+  name: 'users',
+  initialState,
+  reducers: {
+    setUser: (state, action) => {
+      state.user = action.payload
+      localStorage.setItem('userData', JSON.stringify(action.payload))
+    },
+    setLoading: (state, action) => {
+      state.userLoading = action.payload
+    },
+    setError: (state, action) => {
+      state.userError = action.payload
+    },
+  },
+})
+
+export const { setUser, setLoading, setError } = userSlice.actions
+export default userSlice.reducer;
